Using advice-around instead of hook trigger for mode enable.
The original funcition `auto-sudoedit-sudoedit-and-kill' using buffer
redirect method to replace the origin dired displayed buffer with the
tramped one, and indeed that this procedure working for the hook both
for `dired-mode-hook' or `find-file-hook' caused by the original `auto
-sudoedit-mode' trigger. But for thus, the `kill-this-buffer' way will
corrupted the origin buffer argument for the next hook rounder which
will destroyed the whole dired or find-file initializing procedure.

For avoiding this problem, this commit using the around-advices
re-wrappering the `dired' and `find-file' which just transfer the
tramp-transfered path for the rest process powered by the original
package builtin function `auto-sudoedit-tramp-path'.